requirements.txt
Flask
gunicorn
numpy
scikit-learn
pandas
mysql-connector-python
evidently
numpy
PyYAML
-e .